Overhaul transaction control and exception processing
Enables more powerful transactions, better exception handling, and the
option for ExceptionHandler implementations to fail-fast.

Changes to transactions:
* Introduces InitialTransactionBuilder.body(TransactionBody) to replace
transactor(Transactor). TransactionBody accepts a TransactionController,
which can be used for fine-grained control over rollbacks and savepoints
* Deprecates Transactor/RollMeBackException

Changes to exception handling:
* RuntimeExceptions thrown during query execution or a transaction body
are rethrown to the caller after the transaction is rolled back
* Should Connection#rollback throw an exception inside a catch block, it
is added as a suppressed exception where appropriate